refactor: ChatListTemplate Delete
